feat(products): add URLs utilities
The purpose of this module is to provide a single place from which the methods within the `ProductsAPI` get their individual target URLs.

In case, for example, these endpoints change due to external API changes, they can be adjusted within this module, thus avoiding hard-coding the URLs within the ProductsAPI module and ensuring possible future modules get the right values.