refactor(http): move `http` lib to `src/http` dir
This commit is part of a series of upcoming commits
which move libraries within the `packages` directory
to `src/{LIB_NAME}`.

It turned out to be unnecessary to put each module
in it's own library. The implementation for connecting
to the Gelato API has a small footprint anyway.
Thus, it is more appropriate to put all API services
(within `packages/apis/*` directory) and shared
utilities under the `src` directory.