build: uprade node to v18.14.2
We want to move away from `axios` HttpClient to reduce the footprint
of this library. Beginning with Node 18, we are able to use the
built-in Fetch API.

There are some changes we need to make to the current `HttpClient`
in order to migrate to the Fetch API.

For example, the axios client supports generic types for HttpResponses,
which the Fetch API currently lacks.

Using axios, JSON data can be sent directly with HTTP requests.
When using Fetch, however, we explicitly need to JSON.stringify the data.
Similarly, axios internally converts HTTP responses to JSON ,
wheras with Fetch we need to explicitly convert the response by
calling `.json()`.

We also need to keep in mind that we cannot monitor progress of requests
with the Fetch API. This change is trivial for this library, since we don't
intend to provide such a feature.

Lastly, axios considers all requests returning a status code other than 2xx
an error. In Fetch only requests that could not be completed due to any
issue from the target serve's side are considered an error.