refactor(http): move the `HttpClient` to `src/client/http`
The HttpClient should live inside the `src/client` directory
due to the fact that the client is the only class that has
direct access to it.

In addition to this refactor, the HttpClient class extends
a base class `HttpClientBase` which is responsible for
initializing request configurations, such as the request
headers with the `X-API-KEY header.