style: fix endOfLine lint/formatting warnings
Due to differences in how line breaks are handled on Windows & Linux,
the eslint/prettier VSCode plugin complains when the local VSCode
settings do not match the `endOfLine` configuration in the
`.eslintrc.json` file.

To avoid these errors/warnings cross-platform and to auto-detect
the operating system's default behavior, we set the `endOfLine`
setting to `auto` in the `.eslintrc.json` file.

See: https://stackoverflow.com/questions/53516594/why-do-i-keep-getting-eslint-delete-cr-prettier-prettier