fix(orders): use accepted HTTP method to `getOrders`
In commit #f5fb8db we replaced the *GET* with *POST* method to retrieve
a list of orders. This change resulted in an `405` http status code.

This commit reverts this change.

The Gelato API seems to use, although counter-intuitively, a `POST`
HTTP method to *get* a list of orders.

I don't quit understand why you would do that, since it completely
undermines proven practices in designing REST APIs.

Fixes: #38